Leasing (#95)
* Support for leasing in balances and validation components

* Implement storage for leasing

* Move txs IDs to common parts of V1 and V2

These IDs are needed for state validations, and state operates on common structs of v1/v2 transactions.

* Leases state:  store lease's sender as well

It is needed to validate LeaseCancel: only Lease sender can cancel it.

* Leases state: store amount instead of in/out

* Implement validation for Lease and LeaseCancel transactions

* Fixes & first cancellation of leases

* Allow temporary negative balance for diffs of Exchange transactions

* Safe sum function for uint64, tests for both

* Use big.Int for asset quantities, since some reissue transactions overflows even MaxUint64

The constant of 16 bytes is allocated for quantity in asset history record. It's a must, because `history` package
only works with fixed-size records, while big.Int isn't fixed-size.
Anyway, 16 bytes should be enough, since max quantity per one reissue transaction is limited by MaxInt64,
so 2 ** 128 / 2 ** 63 reissue transactions would be needed to overflow it.

* Importer cmd: add profiling

* Use different key prefixes in DB for Waves and asset balances

It's needed to iterate just Waves balances in several cases (leases, waves addresses number etc.).

* Change ResetEffectiveBalanceAtHeight to prover value for TestNet

* Implement all Lease cancellations

* Fix static analyzer's complaints

* Tiny style fix

* Add static analyzer checks to Makefile

* Reduce amount of copy-paste in state validation tests

* Makefile: perform vet and fmt checks before building/testing

* Remove commented code

* Basic tests for leasing

* Add more tests for leasing validations

* Add cancellation-related tests for leases

* Recover panic in Block unmarshal functions

There are no checks for data length in unmarshal functions, so it used to just panic when data was smaller than
expected. It is also impossible to check the length at once, since the proper length is calculated dynamically
depending on data itself, and with multiple checks it seems to slow down too much, so we go for recover anti-pattern.

* Add lease cancellation-related tests for balances

* Implement simple logging for lease cancellation